+++ app/components/sheet-live.vue
@@ -57,7 +57,7 @@
 					>
 						<g class="measure" v-for="(measure, i4) of staff.measures" :key="i4">
 							<g v-for="(token, i5) of measure.matchedTokens" :key="i5"
-								:transform="`translate(${token.x}, ${token.y})` + (token.scale && token.scale !== 1 ? ` scale(${token.scale})` : '')"
+								:transform="`translate(${token.x + token.musicFontNoteOffset}, ${token.y})` + (token.scale && token.scale !== 1 ? ` scale(${token.scale})` : '')"
 								class="token matched"
 							>
 								<text :data-href="token.href">{{token.fontChar}}</text>
@@ -396,6 +396,9 @@
 	@import "../styles/sheetConstants.scss";
 
 
+	$musicFontSize: 2.2px;
+
+
 	.sheet
 	{
 		.mark
@@ -412,7 +415,7 @@
 				text
 				{
 					user-select: none;
-					font-size: 2.2px;
+					font-size: $musicFontSize;
 					fill: $token-default-color;
 
 					&.on
@@ -430,7 +433,7 @@
 			text
 			{
 				user-select: none;
-				font-size: 2.2px;
+				font-size: $musicFontSize;
 			}
 
 			.alter