+++ inc/staffSvg/utils.ts
@@ -10,6 +10,8 @@ const NOTEHEAD_BASE_SCALE = 0.004;
 // 2nd degree chord note head intervals for WHOLE : HALF : SOLID : CROSS = 1.81 : 1.32 : 1.25 : ?
 const NOTE_TYPE_INTERVAL_FACTORS = [1.25 / 1.81, 1.25 / 1.32, 1, 1];
 
+const MUSIC_FONT_NOTE_OFFSETS = [0.05, 0.05, -0.02, 0.1];
+
 const ALTER_WIDTHS = {
 	[-2]: 1.80,
 	[-1]: 1.16,
@@ -23,6 +25,7 @@ const constants = {
 	CLOSED_NOTEHEAD_INTERVAL_FIRST_DEG,
 	NOTE_TYPE_INTERVAL_FACTORS,
 	ALTER_WIDTHS,
+	MUSIC_FONT_NOTE_OFFSETS,
 };