+++ inc/staffSvg/svgSymbols.ts
@@ -134,6 +134,8 @@ const symbolRules = [
 
 	conditionSymbol("NULL LARGE_RECT", elem => elem.identity.type === "rect" && (elem.identity.height > 1) && elem.identity.width >= 2 && elem.rx <= 0 && elem.ry <= 0),
 	conditionSymbol("NULL ENGRAVER_SIG", elem => elem.identity.type === "text" && /www\.lilypond\.org/.test(elem.identity.text)),
+
+	conditionSymbol("TEXT", elem => elem.identity.type === "text"),
 ];
 
 
@@ -173,6 +175,10 @@ const postSymbolRules = [
 
 	postSymbolProcess("NOTEHEAD", (elem, result) => {
 		result.scale = elem.identity.scale.x / NOTEHEAD_BASE_SCALE;
+	}),
+
+	postSymbolProcess("TEXT", (elem, result) => {
+		result.text = elem.identity.text;
 	}),
 ];