+++ app/components/sheet-signs.vue
@@ -23,7 +23,12 @@
 					:points="sign.def.points"
 					:stroke-width="sign.def['stroke-width']"
 				/>
-				<text v-if="sign.def.type === 'text'" :font-size="sign.def['font-size']" :font-weight="sign.def['font-weight']">
+				<text v-if="sign.def.type === 'text'"
+					:font-size="sign.def['font-size']"
+					:font-weight="sign.def['font-weight']"
+					:font-style="sign.def['font-style']"
+					:text-anchor="sign.def['text-anchor']"
+				>
 					<tspan>{{sign.def.text}}</tspan>
 				</text>
 			</g>